consumer/requirements.txt
kafka-python
tweepy
python-twitter